[FEAT] 공개 버튼 드롭다운 구현
