+++ src/components/PostTable.jsx
@@ -1,221 +1,10 @@
-import { React, useState } from "react";
+import { React } from "react";
 import { Link } from "react-router-dom";
-
-function OpenedDropdown() {
-    return(
-        <div className="flex flex-col justify-start items-start">
-  <div className="flex flex-col justify-start items-start flex-grow-0 flex-shrink-0 overflow-hidden gap-2.5 px-4 py-1 rounded-tl-[20px] rounded-tr-[20px] bg-neutral-100/70">
-    <div className="flex justify-start items-center flex-grow-0 flex-shrink-0 relative">
-      <div className="flex justify-start items-start flex-grow-0 flex-shrink-0 relative gap-px">
-        <p className="flex-grow-0 flex-shrink-0 text-base text-left text-[#808593]">비공개</p>
-        <svg
-          width={19}
-          height={20}
-          viewBox="0 0 19 20"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-          className="flex-grow-0 flex-shrink-0 w-[19px] h-[19px]"
-          preserveAspectRatio="xMidYMid meet"
-        >
-          <mask
-            id="mask0_490_9253"
-            style={{ maskType: "alpha" }}
-            maskUnits="userSpaceOnUse"
-            x={0}
-            y={0}
-            width={19}
-            height={20}
-          >
-            <rect y="0.5" width={19} height={19} fill="#D9D9D9" />
-          </mask>
-          <g mask="url(#mask0_490_9253)">
-            <path
-              d="M4.9936 17.5209C4.59878 17.5209 4.26156 17.3811 3.98194 17.1015C3.70231 16.8219 3.5625 16.4846 3.5625 16.0898V8.66039C3.5625 8.26557 3.70231 7.92835 3.98194 7.64873C4.26156 7.3691 4.59878 7.22929 4.9936 7.22929H5.9375V5.64596C5.9375 4.6574 6.2841 3.81652 6.97731 3.12331C7.67054 2.4301 8.51143 2.0835 9.49998 2.0835C10.4885 2.0835 11.3294 2.4301 12.0226 3.12331C12.7159 3.81652 13.0625 4.6574 13.0625 5.64596V7.22929H14.0064C14.4012 7.22929 14.7384 7.3691 15.018 7.64873C15.2976 7.92835 15.4375 8.26557 15.4375 8.66039V16.0898C15.4375 16.4846 15.2976 16.8219 15.018 17.1015C14.7384 17.3811 14.4012 17.5209 14.0064 17.5209H4.9936ZM4.9936 16.3334H14.0064C14.0774 16.3334 14.1358 16.3106 14.1815 16.2649C14.2271 16.2192 14.25 16.1609 14.25 16.0898V8.66039C14.25 8.58933 14.2271 8.53097 14.1815 8.48529C14.1358 8.43961 14.0774 8.41677 14.0064 8.41677H4.9936C4.92254 8.41677 4.86418 8.43961 4.8185 8.48529C4.77282 8.53097 4.74998 8.58933 4.74998 8.66039V16.0898C4.74998 16.1609 4.77282 16.2192 4.8185 16.2649C4.86418 16.3106 4.92254 16.3334 4.9936 16.3334ZM9.49998 13.7605C9.88465 13.7605 10.2117 13.6258 10.4812 13.3563C10.7506 13.0868 10.8854 12.7598 10.8854 12.3751C10.8854 11.9904 10.7506 11.6634 10.4812 11.3939C10.2117 11.1244 9.88465 10.9897 9.49998 10.9897C9.11531 10.9897 8.78825 11.1244 8.51879 11.3939C8.24932 11.6634 8.11458 11.9904 8.11458 12.3751C8.11458 12.7598 8.24932 13.0868 8.51879 13.3563C8.78825 13.6258 9.11531 13.7605 9.49998 13.7605ZM7.12498 7.22929H11.875V5.64596C11.875 4.98623 11.6441 4.42547 11.1823 3.96366C10.7205 3.50186 10.1597 3.27096 9.49998 3.27096C8.84026 3.27096 8.27949 3.50186 7.81769 3.96366C7.35588 4.42547 7.12498 4.98623 7.12498 5.64596V7.22929Z"
-              fill="#808593"
-            />
-          </g>
-        </svg>
-      </div>
-      <svg
-        width={24}
-        height={24}
-        viewBox="0 0 24 24"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-        className="flex-grow-0 flex-shrink-0 w-6 h-6"
-        preserveAspectRatio="xMidYMid meet"
-      >
-        <mask
-          id="mask0_490_11001"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x={0}
-          y={0}
-          width={24}
-          height={24}
-        >
-          <rect x={24} y={24} width={24} height={24} transform="rotate(180 24 24)" fill="#D9D9D9" />
-        </mask>
-        <g mask="url(#mask0_490_11001)">
-          <path
-            d="M12.0003 9.33825C12.1209 9.33825 12.233 9.35748 12.3369 9.39595C12.4407 9.43442 12.5394 9.50044 12.633 9.59403L17.1272 14.0883C17.2657 14.2267 17.3365 14.4007 17.3397 14.6104C17.343 14.82 17.2721 14.9972 17.1272 15.1421C16.9824 15.2869 16.8067 15.3594 16.6003 15.3594C16.3939 15.3594 16.2183 15.2869 16.0734 15.1421L12.0003 11.069L7.92724 15.1421C7.78879 15.2805 7.61476 15.3514 7.40514 15.3546C7.19554 15.3578 7.01831 15.2869 6.87344 15.1421C6.72856 14.9972 6.65611 14.8216 6.65611 14.6152C6.65611 14.4088 6.72856 14.2331 6.87344 14.0883L11.3677 9.59403C11.4612 9.50044 11.56 9.43442 11.6638 9.39595C11.7676 9.35748 11.8798 9.33825 12.0003 9.33825Z"
-            fill="#808593"
-          />
-        </g>
-      </svg>
-    </div>
-  </div>
-  <div
-    className="flex flex-col justify-start items-start flex-grow-0 flex-shrink-0 overflow-hidden rounded-bl-[14px] rounded-br-[14px]"
-    style={{ filter: "drop-shadow(1px 2px 10px rgba(0,0,0,0.03))" }}
-  >
-    <div className="flex justify-start items-start flex-grow-0 flex-shrink-0 relative gap-px px-[21px] py-1.5 bg-white/60 border-t-0 border-r-0 border-b border-l-0 border-[#e7eaf2]">
-      <p className="flex-grow-0 flex-shrink-0 text-base text-left text-[#808593]">전체공개</p>
-      <svg
-        width={19}
-        height={19}
-        viewBox="0 0 19 19"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-        className="flex-grow-0 flex-shrink-0 w-[19px] h-[19px]"
-        preserveAspectRatio="xMidYMid meet"
-      >
-        <mask
-          id="mask0_490_9271"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x={0}
-          y={0}
-          width={19}
-          height={19}
-        >
-          <rect width={19} height={19} fill="#D9D9D9" />
-        </mask>
-        <g mask="url(#mask0_490_9271)">
-          <path
-            d="M9.50128 17.0206C8.47008 17.0206 7.497 16.823 6.58203 16.4276C5.66705 16.0323 4.86904 15.4941 4.18799 14.8131C3.50696 14.1321 2.96879 13.3341 2.57347 12.4191C2.17813 11.5041 1.98047 10.531 1.98047 9.49982C1.98047 8.46051 2.17813 7.48539 2.57347 6.57447C2.96879 5.66355 3.50696 4.86757 4.18799 4.18653C4.86904 3.5055 5.66705 2.96732 6.58203 2.572C7.497 2.17667 8.47008 1.979 9.50128 1.979C10.5406 1.979 11.5157 2.17667 12.4266 2.572C13.3375 2.96732 14.1335 3.5055 14.8146 4.18653C15.4956 4.86757 16.0338 5.66355 16.4291 6.57447C16.8244 7.48539 17.0221 8.46051 17.0221 9.49982C17.0221 10.531 16.8244 11.5041 16.4291 12.4191C16.0338 13.3341 15.4956 14.1321 14.8146 14.8131C14.1335 15.4941 13.3375 16.0323 12.4266 16.4276C11.5157 16.823 10.5406 17.0206 9.50128 17.0206ZM9.50128 15.8164C9.90524 15.2805 10.2453 14.74 10.5213 14.195C10.7974 13.65 11.0222 13.0542 11.1958 12.4076H7.8068C7.99051 13.0745 8.21786 13.6804 8.48886 14.2255C8.75985 14.7705 9.09732 15.3008 9.50128 15.8164ZM7.96822 15.5987C7.66475 15.1633 7.39223 14.6682 7.15067 14.1136C6.9091 13.5589 6.72133 12.9902 6.58735 12.4076H3.90174C4.31992 13.2298 4.88069 13.9205 5.58406 14.4797C6.28744 15.039 7.08216 15.412 7.96822 15.5987ZM11.0343 15.5987C11.9204 15.412 12.7151 15.039 13.4185 14.4797C14.1219 13.9205 14.6826 13.2298 15.1008 12.4076H12.4152C12.2559 12.9953 12.0554 13.5665 11.8138 14.1212C11.5723 14.6759 11.3124 15.1684 11.0343 15.5987ZM3.40393 11.2202H6.34681C6.29708 10.9258 6.26105 10.6373 6.23873 10.3547C6.2164 10.072 6.20524 9.78705 6.20524 9.49982C6.20524 9.21259 6.2164 8.92763 6.23873 8.64496C6.26105 8.36229 6.29708 8.07379 6.34681 7.77945H3.40393C3.32781 8.04842 3.26945 8.32804 3.22885 8.61832C3.18825 8.9086 3.16795 9.20243 3.16795 9.49982C3.16795 9.79721 3.18825 10.091 3.22885 10.3813C3.26945 10.6716 3.32781 10.9512 3.40393 11.2202ZM7.53429 11.2202H11.4683C11.518 10.9258 11.554 10.6399 11.5764 10.3623C11.5987 10.0847 11.6099 9.79721 11.6099 9.49982C11.6099 9.20243 11.5987 8.91494 11.5764 8.63736C11.554 8.35976 11.518 8.07379 11.4683 7.77945H7.53429C7.48456 8.07379 7.44852 8.35976 7.42619 8.63736C7.40386 8.91494 7.3927 9.20243 7.3927 9.49982C7.3927 9.79721 7.40386 10.0847 7.42619 10.3623C7.44852 10.6399 7.48456 10.9258 7.53429 11.2202ZM12.6558 11.2202H15.5986C15.6748 10.9512 15.7331 10.6716 15.7737 10.3813C15.8143 10.091 15.8346 9.79721 15.8346 9.49982C15.8346 9.20243 15.8143 8.9086 15.7737 8.61832C15.7331 8.32804 15.6748 8.04842 15.5986 7.77945H12.6558C12.7055 8.07379 12.7415 8.36229 12.7638 8.64496C12.7862 8.92763 12.7973 9.21259 12.7973 9.49982C12.7973 9.78705 12.7862 10.072 12.7638 10.3547C12.7415 10.6373 12.7055 10.9258 12.6558 11.2202ZM12.4152 6.59199H15.1008C14.6776 5.75971 14.1206 5.06902 13.4299 4.51992C12.7392 3.97083 11.9407 3.5953 11.0343 3.39332C11.3378 3.85411 11.6078 4.35804 11.8443 4.9051C12.0808 5.45217 12.2711 6.01447 12.4152 6.59199ZM7.8068 6.59199H11.1958C11.0121 5.93023 10.7809 5.32049 10.5023 4.76276C10.2237 4.20505 9.89002 3.67853 9.50128 3.18323C9.11255 3.67853 8.77888 4.20505 8.50028 4.76276C8.22167 5.32049 7.99051 5.93023 7.8068 6.59199ZM3.90174 6.59199H6.58735C6.73148 6.01447 6.92179 5.45217 7.15829 4.9051C7.39477 4.35804 7.66475 3.85411 7.96822 3.39332C7.05679 3.5953 6.25699 3.9721 5.56884 4.52374C4.88069 5.07537 4.325 5.76479 3.90174 6.59199Z"
-            fill="#808593"
-          />
-        </g>
-      </svg>
-    </div>
-    <div className="flex justify-start items-start flex-grow-0 flex-shrink-0 relative gap-px px-[21px] py-1.5 bg-white/60">
-      <p className="flex-grow-0 flex-shrink-0 text-base text-left text-[#808593]">친구공개</p>
-      <svg
-        width={19}
-        height={19}
-        viewBox="0 0 19 19"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-        className="flex-grow-0 flex-shrink-0 w-[19px] h-[19px]"
-        preserveAspectRatio="xMidYMid meet"
-      >
-        <mask
-          id="mask0_490_10873"
-          style={{ maskType: "alpha" }}
-          maskUnits="userSpaceOnUse"
-          x={0}
-          y={0}
-          width={19}
-          height={19}
-        >
-          <rect width={19} height={19} fill="#D9D9D9" />
-        </mask>
-        <g mask="url(#mask0_490_10873)">
-          <path
-            d="M9.49998 9.25645C8.73801 9.25645 8.08572 8.98515 7.5431 8.44254C7.00048 7.89992 6.72917 7.24762 6.72917 6.48566C6.72917 5.72368 7.00048 5.07138 7.5431 4.52878C8.08572 3.98615 8.73801 3.71484 9.49998 3.71484C10.2619 3.71484 10.9142 3.98615 11.4569 4.52878C11.9995 5.07138 12.2708 5.72368 12.2708 6.48566C12.2708 7.24762 11.9995 7.89992 11.4569 8.44254C10.9142 8.98515 10.2619 9.25645 9.49998 9.25645ZM3.5625 14.0826V13.5254C3.5625 13.1377 3.6678 12.7786 3.87841 12.4483C4.08901 12.1179 4.3704 11.8639 4.72259 11.6863C5.50512 11.3026 6.29451 11.0149 7.09074 10.8231C7.88697 10.6313 8.69005 10.5353 9.49998 10.5353C10.3099 10.5353 11.113 10.6313 11.9092 10.8231C12.7055 11.0149 13.4948 11.3026 14.2774 11.6863C14.6296 11.8639 14.9109 12.1179 15.1215 12.4483C15.3322 12.7786 15.4375 13.1377 15.4375 13.5254V14.0826C15.4375 14.4165 15.3205 14.7004 15.0865 14.9344C14.8526 15.1683 14.5687 15.2853 14.2348 15.2853H4.7652C4.43129 15.2853 4.14736 15.1683 3.91343 14.9344C3.67948 14.7004 3.5625 14.4165 3.5625 14.0826ZM4.74998 14.0978H14.25V13.5254C14.25 13.365 14.2035 13.2166 14.1107 13.0801C14.0178 12.9436 13.8917 12.8322 13.7323 12.7459C13.0503 12.4099 12.3548 12.1554 11.646 11.9824C10.9371 11.8093 10.2217 11.7228 9.49998 11.7228C8.77822 11.7228 8.06289 11.8093 7.35401 11.9824C6.64512 12.1554 5.94966 12.4099 5.26761 12.7459C5.10826 12.8322 4.98216 12.9436 4.88929 13.0801C4.79642 13.2166 4.74998 13.365 4.74998 13.5254V14.0978ZM9.49998 8.06899C9.9354 8.06899 10.3081 7.91396 10.6182 7.60389C10.9283 7.29382 11.0833 6.92107 11.0833 6.48566C11.0833 6.05024 10.9283 5.6775 10.6182 5.36743C10.3081 5.05736 9.9354 4.90232 9.49998 4.90232C9.06456 4.90232 8.69182 5.05736 8.38175 5.36743C8.07168 5.6775 7.91665 6.05024 7.91665 6.48566C7.91665 6.92107 8.07168 7.29382 8.38175 7.60389C8.69182 7.91396 9.06456 8.06899 9.49998 8.06899Z"
-            fill="#808593"
-          />
-        </g>
-      </svg>
-    </div>
-  </div>
-</div>
-    );
-}
-
-function ClosedDropdown() {
-  return(
-    <div className="flex flex-col justify-start items-start overflow-hidden gap-2.5 px-4 py-1 rounded-[20px] bg-neutral-100 cursor-pointer">
-                                <div className="flex justify-start items-center flex-grow-0 flex-shrink-0 relative">
-                                    <div className="flex justify-start items-start flex-grow-0 flex-shrink-0 relative gap-px">
-                                        <p className="flex-grow-0 flex-shrink-0 text-base text-left text-[#808593]">비공개</p>
-                                        <svg
-                                            width={19}
-                                            height={20}
-                                            viewBox="0 0 19 20"
-                                            fill="none"
-                                            xmlns="http://www.w3.org/2000/svg"
-                                            className="flex-grow-0 flex-shrink-0 w-[19px] h-[19px]"
-                                            preserveAspectRatio="xMidYMid meet"
-                                        >
-                                            <mask
-                                                id="mask0_506_11011"
-                                                style={{ maskType: "alpha" }}
-                                                maskUnits="userSpaceOnUse"
-                                                x={0}
-                                                y={0}
-                                                width={19}
-                                                height={20}
-                                            >
-                                                <rect y="0.5" width={19} height={19} fill="#D9D9D9" />
-                                            </mask>
-                                            <g mask="url(#mask0_506_11011)">
-                                                <path
-                                                    d="M4.9936 17.5209C4.59878 17.5209 4.26156 17.3811 3.98194 17.1015C3.70231 16.8219 3.5625 16.4846 3.5625 16.0898V8.66039C3.5625 8.26557 3.70231 7.92835 3.98194 7.64873C4.26156 7.3691 4.59878 7.22929 4.9936 7.22929H5.9375V5.64596C5.9375 4.6574 6.2841 3.81652 6.97731 3.12331C7.67054 2.4301 8.51143 2.0835 9.49998 2.0835C10.4885 2.0835 11.3294 2.4301 12.0226 3.12331C12.7159 3.81652 13.0625 4.6574 13.0625 5.64596V7.22929H14.0064C14.4012 7.22929 14.7384 7.3691 15.018 7.64873C15.2976 7.92835 15.4375 8.26557 15.4375 8.66039V16.0898C15.4375 16.4846 15.2976 16.8219 15.018 17.1015C14.7384 17.3811 14.4012 17.5209 14.0064 17.5209H4.9936ZM4.9936 16.3334H14.0064C14.0774 16.3334 14.1358 16.3106 14.1815 16.2649C14.2271 16.2192 14.25 16.1609 14.25 16.0898V8.66039C14.25 8.58933 14.2271 8.53097 14.1815 8.48529C14.1358 8.43961 14.0774 8.41677 14.0064 8.41677H4.9936C4.92254 8.41677 4.86418 8.43961 4.8185 8.48529C4.77282 8.53097 4.74998 8.58933 4.74998 8.66039V16.0898C4.74998 16.1609 4.77282 16.2192 4.8185 16.2649C4.86418 16.3106 4.92254 16.3334 4.9936 16.3334ZM9.49998 13.7605C9.88465 13.7605 10.2117 13.6258 10.4812 13.3563C10.7506 13.0868 10.8854 12.7598 10.8854 12.3751C10.8854 11.9904 10.7506 11.6634 10.4812 11.3939C10.2117 11.1244 9.88465 10.9897 9.49998 10.9897C9.11531 10.9897 8.78825 11.1244 8.51879 11.3939C8.24932 11.6634 8.11458 11.9904 8.11458 12.3751C8.11458 12.7598 8.24932 13.0868 8.51879 13.3563C8.78825 13.6258 9.11531 13.7605 9.49998 13.7605ZM7.12498 7.22929H11.875V5.64596C11.875 4.98623 11.6441 4.42547 11.1823 3.96366C10.7205 3.50186 10.1597 3.27096 9.49998 3.27096C8.84026 3.27096 8.27949 3.50186 7.81769 3.96366C7.35588 4.42547 7.12498 4.98623 7.12498 5.64596V7.22929Z"
-                                                    fill="#808593"
-                                                />
-                                            </g>
-                                        </svg>
-                                    </div>
-                                    <svg
-                                        width={24}
-                                        height={24}
-                                        viewBox="0 0 24 24"
-                                        fill="none"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                        className="flex-grow-0 flex-shrink-0 w-6 h-6"
-                                        preserveAspectRatio="xMidYMid meet"
-                                    >
-                                        <mask
-                                            id="mask0_506_11014"
-                                            style={{ maskType: "alpha" }}
-                                            maskUnits="userSpaceOnUse"
-                                            x={0}
-                                            y={0}
-                                            width={24}
-                                            height={24}
-                                        >
-                                            <rect width={24} height={24} fill="#D9D9D9" />
-                                        </mask>
-                                        <g mask="url(#mask0_506_11014)">
-                                            <path
-                                                d="M11.9997 14.6617C11.8791 14.6617 11.767 14.6425 11.6631 14.604C11.5593 14.5656 11.4606 14.4996 11.367 14.406L6.87276 9.91175C6.73429 9.7733 6.66346 9.59927 6.66026 9.38965C6.65704 9.18003 6.72788 9.00279 6.87276 8.85792C7.01763 8.71306 7.19326 8.64062 7.39966 8.64062C7.60606 8.64062 7.78169 8.71306 7.92656 8.85792L11.9997 12.931L16.0728 8.85792C16.2112 8.71947 16.3852 8.64864 16.5949 8.64542C16.8045 8.64222 16.9817 8.71306 17.1266 8.85792C17.2714 9.00279 17.3439 9.17843 17.3439 9.38485C17.3439 9.59125 17.2714 9.76688 17.1266 9.91175L12.6323 14.406C12.5388 14.4996 12.44 14.5656 12.3362 14.604C12.2324 14.6425 12.1202 14.6617 11.9997 14.6617Z"
-                                                fill="#808593"
-                                            />
-                                        </g>
-                                    </svg>
-                                </div>
-                            </div>
-  );
-}
-
-function AccessButton() {
-        const [isClicked, setIsClicked] = useState(true);
-
-        const handleClick = () => {
-          setIsClicked(!isClicked);
-        };
-
-        return(
-          isClicked ? <ClosedDropdown onClick={handleClick} /> : <OpenedDropdown onClick={handleClick}/>
-        );
-
-}
+import PrivacyButton from "./Button/PrivacyButton";
 
 function PostPreview( {title, comments, date, author, likes, views} ) {
     return(
-        <div className="flex-grow-0 flex-shrink-0 w-[953px] h-[71px] relative overflow-hidden bg-white hover:bg-neutral-100">
+        <div className="flex-grow-0 flex-shrink-0 w-[953px] h-[71px] relative bg-white hover:bg-neutral-100">
             <svg
               width={953}
               height={1}
@@ -288,12 +77,12 @@ export default function PostTable({ tablename }) {
     return(
         <div className="flex flex-col justify-start items-start flex-grow-0 flex-shrink-0 relative">
         <div className="flex-grow-0 flex-shrink-0 w-[953px] h-[69px]">
-          <div className="w-[953px] absolute left-0 top-0 overflow-hidden gap-2.5 pl-8 pr-[34px] py-5 rounded-tl-[14px] rounded-tr-[14px] bg-neutral-100">
+          <div className="w-[953px] absolute left-0 top-0 gap-2.5 pl-8 pr-[34px] py-5 rounded-tl-[14px] rounded-tr-[14px] bg-neutral-100">
             <div className="flex flex-nowrap w-[891px] h-[29px] relative">
               <p className="h-[29px] text-xl font-medium text-left text-[#414244]">
                 {tablename}
               </p>
-              <AccessButton />
+              <PrivacyButton />
               <Link to='/'>
               <div className="flex flex-nowrap items-center absolute right-[0px] gap-1">
                 <p className="text-lg text-left text-[#808593]">
